dm writecache: fix format string warning
The return type of ACCESS_ONCE is configuration dependent and may be either
'int' or 'long int' for the writecache_has_error() macro, so we get a warning
like this for either format string:

In file included from drivers/md/dm-writecache.c:8:
drivers/md/dm-writecache.c: In function 'writecache_status':
drivers/md/dm-writecache.c:2227:10: error: format '%ld' expects argument of type 'long int', but argument 4 has type 'int' [-Werror=format=]
   DMEMIT("%ld %llu %llu %llu", writecache_has_error(wc),
          ^~~~~~~~~~~~~~~~~~~~
include/linux/device-mapper.h:549:46: note: in definition of macro 'DMEMIT'
      0 : scnprintf(result + sz, maxlen - sz, x))
                                              ^
The code is otherwise correct, so we just need to shut up the warning,
which can be done using an extra type cast.

Fixes: 2105231db61b ("dm: add writecache target")

+++ drivers/md/dm-writecache.c
@@ -2232,7 +2232,7 @@ static void writecache_status(struct dm_target *ti, status_type_t type,
 
 	switch (type) {
 	case STATUSTYPE_INFO:
-		DMEMIT("%ld %llu %llu %llu", writecache_has_error(wc),
+		DMEMIT("%ld %llu %llu %llu", (long)writecache_has_error(wc),
 		       (unsigned long long)wc->n_blocks, (unsigned long long)wc->freelist_size,
 		       (unsigned long long)wc->writeback_size);
 		break;
